requirements/base.txt
Django==5.0.1
djangorestframework==3.16.0
redis>=4.6
psycopg2-binary==2.9.9
django-unfold==0.20.1
requests==2.31.0
daphne==4.1.0
pyTelegramBotAPI==4.26.0
whitenoise==6.9.0
telethon==1.39.0
selenium==4.29.0
Pillow==11.1.0
django-environ==0.11.2
django-celery-beat==2.1.0
celery==5.3.6
redis>=4.6
django-redis==5.2.0
channels==4.0.0
opencv-python==4.11.0.86
pandas==2.2.3
openpyxl==3.1.5